+++ solum/include/solum/value.h
@@ -15,6 +15,7 @@
 typedef struct SolObject SolObject;
 typedef struct SolBlock  SolBlock;
 typedef struct SolArray  SolArray;
+typedef struct SolString SolString;
 
 typedef enum {
     SOL_NIL,
@@ -23,6 +24,7 @@ typedef enum {
     SOL_FLOAT,    /* 45              */
     SOL_BLOCK,    /* { ... }         */
     SOL_ARRAY,    /* [#1, #2]        */
+    SOL_STRING,   /* "hello"         */
     SOL_OBJ
 } SolValueType;
 
@@ -34,6 +36,7 @@ typedef struct {
         double    real;
         SolBlock  *block;
         SolArray  *array;
+        SolString *string;
         SolObject *obj;
     } as;
 } SolValue;
@@ -42,6 +45,7 @@ typedef struct {
 #define SOL_BOOL_VAL(b)   ((SolValue){ SOL_BOOL,  { .boolean = (b) } })
 #define SOL_BLOCK_VAL(b)  ((SolValue){ SOL_BLOCK, { .block = (b) } })
 #define SOL_ARRAY_VAL(a)  ((SolValue){ SOL_ARRAY, { .array = (a) } })
+#define SOL_STRING_VAL(s) ((SolValue){ SOL_STRING, { .string = (s) } })
 #define SOL_INT_VAL(i)    ((SolValue){ SOL_INT,   { .integer = (i) } })
 #define SOL_FLOAT_VAL(f)  ((SolValue){ SOL_FLOAT, { .real = (f) } })
 #define SOL_OBJ_VAL(o)    ((SolValue){ SOL_OBJ,   { .obj = (o) } })
@@ -50,6 +54,7 @@ typedef struct {
 #define SOL_IS_BOOL(v)    ((v).type == SOL_BOOL)
 #define SOL_IS_BLOCK(v)   ((v).type == SOL_BLOCK)
 #define SOL_IS_ARRAY(v)   ((v).type == SOL_ARRAY)
+#define SOL_IS_STRING(v)  ((v).type == SOL_STRING)
 #define SOL_IS_INT(v)     ((v).type == SOL_INT)
 #define SOL_IS_FLOAT(v)   ((v).type == SOL_FLOAT)
 #define SOL_IS_OBJ(v)     ((v).type == SOL_OBJ)
@@ -57,6 +62,7 @@ typedef struct {
 #define SOL_AS_BOOL(v)    ((v).as.boolean)
 #define SOL_AS_BLOCK(v)   ((v).as.block)
 #define SOL_AS_ARRAY(v)   ((v).as.array)
+#define SOL_AS_STRING(v)  ((v).as.string)
 #define SOL_AS_INT(v)     ((v).as.integer)
 #define SOL_AS_FLOAT(v)   ((v).as.real)
 #define SOL_AS_OBJ(v)     ((v).as.obj)